xfstests: remove date from expected tarball name
The expected tarball name is xfstests-20110701 so we'd have to update
the shellpack any time we'd like to update the tarball. As xfstests do
not do any point releases and using the latest git is common, change the
base tarball name to xfstests-git. This needs an update on the mirrors.

Signed-off-by: David Sterba <dsterba@suse.com>
Signed-off-by: Mel Gorman <mgorman@suse.com>