pgioperf: Flush output after each printed stat
Currently pgioperf leaves flushing of printf() calls on glibc which
actually caches it rather aggresively as the writes are redirected to a
file. As a result output from different processes is not interleaved as
expected which makes it somewhat more difficult to guess interaction
between reader and writer processes. fflush() output after each written
statistics to make interleaving of output lines at least roughly match
the real time when statistics was computed.

More exact solution would be to timestamp output lines but for
occasional manual analysis it seems like an overkill.

Signed-off-by: Jan Kara <jack@suse.cz>