fio: Bump version to 2.20
Also remove patch that fixed a double unlock. The issue was fixed in mainline
fio with commit 42e833f ("mutex: fix double unlock in
fio_mutex_down_timeout()").

Note that starting with fio-2.13 bandwidth stats information was not displayed
-- caused by commit 9e4438f ("stat: don't trust per_unit_log() if log is
NULL"). This was fixed with 3718170 ("stat: Re-add output of basic bw
information if bw_log is not written") which is included in fio-2.20.

We use this bandwidth data when extracting fio benchmarks results. Thus
unpatched fio versions between 2.13-2.19 can't be used in our environment.

Signed-off-by: Andreas Herrmann <aherrmann@suse.com>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>