run-kvm: Add option to skip installing host kernel in guest
An initrd image is created at the end of a SUSE Linux Enterprise
installation process, incorporating the list of currently loaded
modules.

Initrd images are not always compatible between bare metal and VM. For
example, the virtio_blk module is likely to be loaded in the VM but not
on bare metal. Which means that copying the host's initrd image to a
guest as part of the run-kvm.sh script can leave the guest unable to
find the root disk.

This isn't usually a problem if you build custom kernels and include the
virto-blk driver. But it is definitely a problem when using the
official SLE kernels and initrd files.

Allow the user to keep the guest kernel and initrd as-is by using the
--keep-kernel option.

It's a new option and not the default behaviour for two reasons:

  1. It would break existing workflows

  2. Unless you know what you're doing, you should keep the host and
     guest kernels in sync

Signed-off-by: Matt Fleming <matt@codeblueprint.co.uk>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>