compare-mmtests: compress JSON exports
Make compare-mmtests.pl compress JSON exports. This change prepares the
field for a forthcoming patch that makes cache-mmtests.sh consider JSON
exports too, in addition to caching standard output.

A common use case for JSON exports is to be used by some external reporting
infrastructure, which stores them compressed. Before this change,
compression was performed by format-for-display.sh.

As the caching layer is introduced, the exports will have to, necessarily,
be stored compressed in a cache directory. If caching had to transparently
reproduce the current compare-mmtests.pl behaviour this would imply that in
the event of a cache hit the files would be uncompressed -- just to be
re-compressed immediately afterward by the reporting infrastructure.

By modifying the user expectation that JSON's come out uncompressed we allow
the cache to avoid wasteful work in case of a hit.

Signed-off-by: Giovanni Gherdovich <ggherdovich@suse.cz>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>