cache-mmtests: handle JSON exports
Cache JSON reports in addition to standard output.

1. DESIGN
2. QUIRKS
3. TEST RESULTS

1. DESIGN

Caching of JSON exports can be implemented in two ways: keeping the file in
double copy (in the benchmark directory and in the cache), or using a
symbolic link to the cached files. We opt for the symbolic linking strategy
in order to save space. As a bonus, this avoids copying them from the cache
to the script invocation directory and vice-versa.

Using symbolic links in this way may be prone to race conditions. This can
happen in two ways:

1. the target file is deleted and the link goes dead
2. the target file is replaced with a different file

(2) would be the most concerning as it would give misleading data in
reports, but is very unlikely to happen as the hash of a cache entry
depends on both the compare-mmtests invocation command and the benchmark
directory: if a JSON report is there, it must be the right one.

Case (1) can actually happen if periodical eviction of old entries is
performed, leading to missing data (which is not as bad as wrong data). The
timing of cache trimming within our internal infrastructure makes this
condition unlikely to verify, so we don't address this case.

2. QUIRKS

If the cache entry $CACHEDIR exists but the file are unusable (eg. missing or
corrupt) the JSON export is saved in the benchmark directory as a regular
file, and not as a symbolic link to a cache entry. At the following execution
of compare-mmtests the nominal behavior is restored.

3. TEST RESULTS

The test script is omitted as it depends on specific data present on our
machine. The "OUTPUT" column refers to the comparison with the expected
output of the script; the "CACHE" column refers to checking for the integrity
of the cache if the script is functioning nominally, or checking if the
cache is effectively cleaned up if the script is functioning in degraded mode.
The "JSON EXPORT" column corresponds to checking if the JSON export is a
symbolic link (nominal behavior) or a regular file (in case of expected
degraded behaviour).

    DESCRIPTION                                       OUTPUT  CACHE   JSON EXPORT
    -----------------------------------------------------------------------------
    COMPARE-MMTESTS W/ JSON EXPORT
    Nominal: cache miss                               PASS    PASS    PASS
    Nominal: cache hit                                PASS    PASS    PASS
    Degraded: missing cache.gz                        PASS    PASS    PASS
    Degraded: missing cache.json.gz                   PASS    PASS    PASS
    Degraded: both cache{.gz, .json.gz} missing       PASS    PASS    PASS
    Degraded: corrupt cache.gz                        PASS    PASS    PASS
    Degraded: corrupt cache.json.gz                   PASS    PASS    PASS
    Degraded: both cache{.gz, .json.gz} corrupt       PASS    PASS    PASS

    COMPARE-MMTESTS W/O JSON EXPORT
    Nominal: compare-mmtests w/o json export          PASS    PASS    -
    Degraded: compare-mmtests, no json, missing cache PASS    PASS    -
    Degraded: compare-mmtests, no json, corrupt cache PASS    PASS    -

    EXTRACT-MMTESTS
    Nominal: extract-mmtests                          PASS    PASS    -
    Degraded: extract-mmtests, missing cache          PASS    PASS    -
    Degraded: extract-mmtests, corrupt cache          PASS    PASS    -

Signed-off-by: Giovanni Gherdovich <ggherdovich@suse.cz>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>