bonnie++: Increase number of files for directory benchmarks
Delete runs of a bonnie++ benchmarks are extremely fast - under 1 second
on some machines - which is too short for an IO benchmark and thus is
prone to disturbance by random effects. Modify bonnie configuration to
create 20000 files instead of 1024 to make delete runs longer. Note that
this does not influence create & read time numbers much since the total
size of created files (which dominates the runtime) is kept the same.

Signed-off-by: Jan Kara <jack@suse.cz>