bonnie++: Reduce number of iterations for async runs
Currently we do 10 iterations of async bonnie++ runs to capture better
variation of the delete runs of the benchmark. Since we have increased
number of files now, delete runs look reasonably stable so there's no
need for that many runs (which are quite time consuming). Reduce the
number of iterations to 5.

Signed-off-by: Jan Kara <jack@suse.cz>