mmtests: Do not use TESTDISK for benchmark compilation
Currently mmtests use SHELLPACK_TEST_MOUNT (where TESTDISK gets mounted)
for extracting benchmark sources and the compilation of the benchmark.
This makes it impossible to recreate filesystem on TESTDISK between
benchmark iterations and also creates some uncertainty in the initial
state of the filesystem. Configure mmtests so that we use work/ instead
of work/testdisk/ for these purposes.

Signed-off-by: Jan Kara <jack@suse.cz>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>