mmtests: Remove TESTDISK_DIR and TESTDISK_DIRS setting
TESTDISK_DIRS are now completely unused. Remove them. TESTDISK_DIR is
now just an input for run_mmtests.sh where special data directory is set
up. So there's no need to export it or to set it if it is not already
set. Individual tests should not care about this variable anymore.

Signed-off-by: Jan Kara <jack@suse.cz>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>