mmtests: Avoid clearing TESTDISK_FILESYSTEM on cache setup
When using dm-cache, we've created filesystem on target device early,
presumably so that we don't pollute dm-cache with mkfs data. To not
recreate the filesystem again later, we've the set TESTDISK_FILESYSTEM
to "". This however breaks a few tests later in run-mmtests.sh and also
e.g. xfstests-bench.

Fix the problem by going trough normal fs creation after cache device is
set up and then just flush the cache before doing real testing. Note
that bcache would need flushing as well but we did bother with this
previously either so we are bug-to-bug compatible for now.

Signed-off-by: Jan Kara <jack@suse.cz>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>