run-mmtests.sh: Wait for ntp to synchronize time before proceeding
When NTP messes with system time while a benchmark is running it can
confuse the benchmark (e.g. some operation is reported to take extra
long). Wait for NTP to synchronize system clock before starting testing
to avoid this.

The annoying part is that we have to wait couple of minutes before each
test starts. Also note that we leave ntpd running so it could later
still modify system time if it drifts. But hopefully those modifications
will be very small.

Signed-off-by: Jan Kara <jack@suse.cz>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>