dbench4: Sort by number of clients first
Extraction of dbench4 times sorted the resulting array by the
time when the sample has been taken. However the time is relative to the
start of the benchmarking run and thus results of runs for different
numbers of clients got mixed in _ResultData array. Apparently this does
not confuse any of the plotting or comparison functions but it is
definitely confusing when manually inspecting the data. Sort the array
first by a number of clients and use time only as a secondary key.

Signed-off-by: Jan Kara <jack@suse.cz>