compare: Rename _ResultsStdDevTable to _ResultsNormalizedTable
_ResultsStdDevTable is only loosely related to standard deviations. In
fact it is a table of results produced by sdiff() function which
provides distance between means normalized by standard deviations.
Later, when we add other methods of comparing results, the name would be
even more confusing so rename it to _ResultsNormalizedTable which better
describes what it is about.

Also rename related local variables.

Signed-off-by: Jan Kara <jack@suse.cz>