Stat: Factor out selection of data for means
Currently data for a mean of highest/lowest/trimmed subselection is
always computed in the function computing the mean. Factor out the
function doing subselection. As result we fix inconsistency in what
'percentage' argument means - for 'lowest' it was from how big percentage
of data mean should be computed, for 'highest' and 'trimmed' it was how
many data should be removed. We settle on 'percentage' meaning
percentage of data to keep.

Signed-off-by: Jan Kara <jack@suse.cz>