mmtest: Infer mean function from mean name
Name of the function for mean computation is determined by the name of
the mean to compute. When using R to analyze data we will also need to
propagate information about mean to use and mean name is a natural
choice. So make _MeanName the primary source of information and deduce
perl function to compute mean from that.

Signed-off-by: Jan Kara <jack@suse.cz>