mmtests: Add SummariseSubselection
Add new summarization module which reports results using new
subselection mean and confidence interval. It uses R to compute the
values.

Although the algorithm could be reimplemented relatively easily in Perl,
we use R to have a model of Perl-R interaction and to be able to easily
experiment with different algorithm modifications in R. When we settle
on the final algorithm, we can consider Perl implementation for speed...

We also add support to Compare.pm to use different methods for
comparison of results in the ratio table when including standard
deviations / confidence interval lengths. Modules can use
_RatioCompareOp to set special comparison operation, we default to
traditional 'sdiff'.

Signed-off-by: Jan Kara <jack@suse.cz>