ExtractSstartup: Fix sorting of self->{_Operations}
The generated summary matrix of extracted results was different for
baseline and comparison kernels because values in self->{_Operations}
were not sorted.

Also when result files were missing for some test patterns, e.g. for
5r5w-seq case, generated summary matrices could be different for
baseline and comparison kernels.

In both cases matrix entries with different meaning were compared.

Fix sorting (use string comparison).

Fix handling in case a result file is missing. In this case add dummy
ops which cause display of "0.00 (0.00%) "or "<value> (nan%)" in the
comparison table for missing results resp. comparison with missing
results. This change implies that whenever the set of test patterns is
changed in S-startup the array of patterns in this script needs to be
adapted accordingly.

Signed-off-by: Andreas Herrmann <aherrmann@suse.com>