SummariseVariabletime: Push "NaN" into summary for ops without data
This allows to properly compare results for benchmarks where result
data was only partially collected. To use this the corresponding
extract module for a benchmark always needs to push the same set of
ops (even if no data was extracted for an op).

This way Compare.pm can pass these NaN-s into comparison table to
indicate that some benchmark results are missing.

Signed-off-by: Andreas Herrmann <aherrmann@suse.com>