SummariseSingleops: add field _Operations (list of ops)
Add the _Operations field, which is not used but is in preparation for a
change that converts $self->{_SummaryData} from a list to a map: the list of
operation (specifically: their order for display in the table by
compare-mmtests.pl with and without the --print-ratio flag) is essential as
_SummaryData will not preserve the order of operations anymore. The classes
SummariseMultiops and SummariseVariabletime already have the _Operation field.

The _Operations fields is obtained taking the first column of the _ResultData
table.

This is what _ResultData looks like for a SummariseSingleops (specifically,
ExtractSpecjvm):

    [
      ["compiler.compiler", 5266.08],
      ["compiler.sunflow", 2875.92],
      ["compress", "4130.10"],
      ["crypto.aes", "2665.80"],
      ["crypto.rsa", 1839.63],
      ["crypto.signverify", 6667.05],
    ]

In that case, _Operations would be:

  [ "compiler.compiler", "compiler.sunflow", "compress", "crypto.aes", "crypto.rsa", "crypto.signverify" ]

i.e. the first column of _ResultData.

Signed-off-by: Giovanni Gherdovich <ggherdovich@suse.cz>