remove CompareCputime.pm, use CompareMonitor.pm instead
compare-mmtests.pl uses CompareCputime (which is a subclass of Compare).
We're about to radically change the implementation of Compare, but we don't
want to worry about keeping compatibility with monitor data (at least for the
moment).

Change compare-mmtests.pl to use CompareMonitor instead (unused at this
time). CompareMonitor is also a (trivial) subclass of Compare, so the
behaviour is unchanged. In a following patch we'll copy/paste all of the
current code from Compare into CompareMonitor.

We'll then be able to change Compare without affecting monitors. At a later
time we can think about adapting monitors to use the new, re-designed Compare
class.

Signed-off-by: Giovanni Gherdovich <ggherdovich@suse.cz>