compare: Add support for results significance detection
MMtests has a lot of tests. Which is great because it gives a high
degree of confidence that you're testing all scenarios and corner
cases, but also suffers from the problem that reading all of those
results manually is time-consuming and error-prone.

Tests are rarely deleted, but new ones are added from time to time,
which means this human bottleneck is likely to get worse over time.

What is needed is a method to separate the wheat from the chaff, and
for developers to conserve their valuable mental capacity and use it
only when experience and intuition in analysing results is required.

This patch implements a very basic foundation for using statistical
methods to detect when a test result differs significantly from the
baseline using a two-sample t-test. The limitations of this method are
discussed below under "Known Limitations".

"significance" is defined on a per-test basis through the
'_SignificanceLevel' field. By default, if none is specified it's 10%
but that's purely because that's the value I found to work well with
hackbench. Lower values will detect smaller changes in the results,
higher values will only flag large deviations from the baseline as
being interesting.

Significance is intrinsically a benchmark-specific attribute because
it depends on how precisely the benchmark can measure operations. You
can expect micro-benchmarks to have lower significance levels than
macro-benchmarks, since macro-benchmarks necessarily encounter
interrupts/noise from the system due to their size.

The significance detection does not distinguish between a performance
gain or a performance loss when compared with the baseline, so it's
worth bearing in mind that this patch does not just implement
regression detection, it also implements improvement detection. It's
down to the user to understand which has been occurred.

Known Limitations:

The benchmark result data must supply an arithmetic mean to work with
the t-test.

As pointed out by Jan Kara, the t-test suffers from a very basic
dependence on the benchmark result data following a normal
distribution, and this is almost certainly not the case for disk and
file system benchmarks.

A basic next step could be implementing a Mann-Whitney U Test which
does not assume the data has a normal distribution. For bonus points,
cluster analysis might be more useful for those data results that are
bi-modal, i.e. all tests that potentially hit/miss OS caches.

Example:

Here's the output of hackbench results from my 4-cpu Intel Core i7
laptop. The 'slow' column was run with everything pinned to only 2
cpus. 'fast' and 'fast2' are normal out-of-the-mmtests-box runs:

  $ compare-kernels.sh --auto-detect

                            fast                  slow                 fast2
  Amean    1       0.4263 (  0.00%)      0.2091 ( 50.94%)      0.4697 (-10.19%)
  Amean    3       0.5000 (  0.00%)      0.5473 ( -9.46%)      0.6211 (-24.23%)
  Amean    5       0.7841 (  0.00%)      1.3609 (-73.55%)      0.8546 ( -8.98%)
  Amean    7       0.9230 (  0.00%)      1.4947 (-61.94%)      0.9996 ( -8.30%)
  Amean    12      1.3830 (  0.00%)      2.1731 (-57.13%)      1.4030 ( -1.45%)
  Amean    16      1.8286 (  0.00%)      3.0337 (-65.91%)      1.7246 (  5.69%)
  CoeffVar 1      11.5119 (  0.00%)      3.9464 ( 65.72%)     12.0473 ( -4.65%)
  CoeffVar 3      16.4955 (  0.00%)      5.6044 ( 66.02%)     10.7566 ( 34.79%)
  CoeffVar 5      10.1071 (  0.00%)     42.2010 (-317.54%)     12.3308 (-22.00%)
  CoeffVar 7      16.5349 (  0.00%)     12.3487 ( 25.32%)      8.3048 ( 49.77%)
  CoeffVar 12     13.2654 (  0.00%)     11.0402 ( 16.77%)      5.5198 ( 58.39%)
  CoeffVar 16      8.3660 (  0.00%)      7.2666 ( 13.14%)      5.1533 ( 38.40%)
  Significant 1       0.0000 (  0.00%)      1.0000 (  0.00%)      0.0000 ( 13.49%)
  Significant 3       0.0000 (  0.00%)      0.0000 ( 19.81%)      1.0000 (  0.81%)
  Significant 5       0.0000 (  0.00%)      1.0000 (  0.85%)      0.0000 ( 16.45%)
  Significant 7       0.0000 (  0.00%)      1.0000 (  0.00%)      0.0000 ( 27.06%)
  Significant 12      0.0000 (  0.00%)      1.0000 (  0.00%)      0.0000 ( 79.72%)
  Significant 16      0.0000 (  0.00%)      1.0000 (  0.00%)      0.0000 ( 15.09%)

The Significant rows include a 0 or 1 for "null hypothesis rejected",
the null hypothesis being "The performance is the same". The value in
the brackets is the P-value as a percentage, which basically says:

  "What is the chance that the this difference is *NOT* significant".

Cc: Jan Kara <jack@suse.cz>
Cc: Giovanni Gherdovich <ggherdovich@suse.cz>
Signed-off-by: Matt Fleming <matt@codeblueprint.co.uk>
Signed-off-by: Giovanni Gherdovich <ggherdovich@suse.cz>