git_fetch: Support commit different from HEAD of branch master
Treat version number as commit/tag ID. Use it for git-checkout after
repository is cloned and for archiving files from the repository.
If version number is 0 fall back on master.

Note that fetching of subtrees, e.g. using something like
v4.17:tools/perf is not supported. git-archive would support this but
git-checkout does not.

Signed-off-by: Andreas Herrmann <aherrmann@suse.com>