git_fetch: Check whether commit is valid otherwise use master
We pass shellpack version to git_fetch and try to treat it as a
commit/tag ID to checkout/archive the corresponding source code tree.
In case no version is passed or version equals 0 we already use master
as default.

In some cases version is not equal to 0 but still not a valid
commit/tag ID. E.g. in case of dbench4 version is "4.1alpha" but there
is no such commit/tag in its git-tree. Thus we need to verify that
shellpack version really identifies a commit and thus can savely
treated as commit/tag ID.

Add function to check validity of a commit and fall back on master if
its invalid.

Signed-off-by: Andreas Herrmann <aherrmann@suse.com>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>