MonitorTurbostat: Simplify output parsing
The parsing code gets data only from the turbostat summary line. So
simplify the parsing to take that into account and just push gathered
data to _ResultData instead of building array of values to push which we
later eventually push. This also fixes a small glitch that the last
recorded turbostat value wasn't getting inserted in _ResultData.

Signed-off-by: Jan Kara <jack@suse.cz>