ExtractSpecjbbpeak: Drop
The extract module has been broken for a few year at least for
multi-instance case since the file does not exist in the benchmark
output. Mel says this is not really useful so not worth fixing and it is
a very weird instance of SummariseSingleops and with unusual _ResultData
format.

Signed-off-by: Jan Kara <jack@suse.cz>