MonitorMmtestsvmstat: Use standard _ResultData format
Use the same format for _ResultData as consumed by SummariseMultiops.
This will allow generic support for multiple iterations and also
simplifies the code.

Note that this somewhat changes the look of comparison for this monitor.
It will now look like:

                                4.12.14        4.12.14
                        0.gfaef795-rpm-4.12.14-25.22-2200.gd2d8b17-rpm-4.12.14-25.25-222
Minor Faults               107427594.00   104436308.00
Major Faults                       9.00           9.00
Swap Ins                           0.00           0.00
Swap Outs                          0.00           0.00
...

So the values are all floats.

Signed-off-by: Jan Kara <jack@suse.cz>