Summarise: Unify selection of ratio operations
To unify handling of ratio tables between SummariseSingleops and
SummariseMultiops and also to make it similar to the handling of
standard result tables, define new _RatioOperations field. This is
analog of _Operations field for standard result tables. We also convert
all benchmarks using _SingleInclude or _MultiInclude to use this field
instead.

Signed-off-by: Jan Kara <jack@suse.cz>