Extract: Sort resulting data automatically when needed
Some extract modules sort _ResultData by time. This is actually only
needed for plotting and there's no need to handle this inside individual
extract functions. Now that we have common _ResultData format, we can
handle sorting automatically and do that only if we are going to plot
the data. This gets rid of last two extract modules modifying
_ResultData by hand.

Signed-off-by: Jan Kara <jack@suse.cz>