Compare: Determine comparison operation based on compared statistics
Currently each summarisation module defined a set of compare operations.
These operations are in fact determined by the statistic that is being
compared and the type of data being compared. So add to each
summarization module _SummaryStats array that defines a set of
statistics used for summarization (in later patches we make use of this
array also for computing all the statistics) and determine appropriate
comparison function based on this array.

This also fixes a problem where SummariseSubselection was using pdiff
for confidence-interval length comparisons in one case despite the
fact that shorter confidence interval is always better.

Signed-off-by: Jan Kara <jack@suse.cz>