Extract: Generate list of operations
If list of operations is not provided in _Operations, generate it
automatically from data in _ResultData. We use the ordering in which
data was added to the _ResultData array in a hope that that is the most
sensible ordering (definitely it is the one currently used e.g. for
monitors).

Signed-off-by: Jan Kara <jack@suse.cz>