MonitorLatency: Convert to Summarise
Convert MonitorLatency to use Summarise for summarization and plotting.
We drop reporting of sum & avg of samples above threshold as we don't
currently have statistics functions implemented for that and this seems
to be the only place doing such thing. But if needed, necessary
functions can be easily added later.

Signed-off-by: Jan Kara <jack@suse.cz>