Switch to using activity file for result detection
tests-timestamp- files serve more as a general log about what the test
is doing and what is the state of the system. Thus when adding iteration
support, they will naturally belong to a per-iteration data.
tests-activity- is more a high-level log how test is progressing and
thus it makes sense to have this common for all test iterations. Switch
detection whether a directory is result directory and what benchmark is
running to use activity files instead of timestamp files.

Signed-off-by: Jan Kara <jack@suse.cz>