Switch SHELLPACK_LOG to per RUNNAME directory
Switch SHELLPACK_LOG to be SHELLPACK_LOG_BASE/RUNNAME so that each test
run gets recorded in its own directory. That way log data from different
tests are not mixed and we don't have to append RUNNAME to each of them.
It would get even more confusing once we add iteration support to
mmtests.

Also move cleaning of old data after log partition is set up and mounted
as otherwise it has no effect if log partition is in use.

Signed-off-by: Jan Kara <jack@suse.cz>