Remove profile knowledge from extractReport
Every benchmark's extractReport function just used the $profile
argument to append it after $reportDir and there's not much more that
could be done with it. So just remove this argument and do the append in
the two calling sites instead.

Replacement has been done with:

for f in bin/lib/MMTests/Extract*.pm; do
	sed -i -e 's|\$reportDir/\$profile|$reportDir|g' $f
        sed -i -e 's|,[ ]*\$profile)|)|' $f
done

and a handful of manual fixups. In particular SStartup handler has been
hardcoding 'noprofile' as the profile name.

Signed-off-by: Jan Kara <jack@suse.cz>