Iteration support
Add support to run-mmtests.sh to iterate a test several times. The
number of iterations is controlled by MMTEST_ITERATIONS variable in test
configuration and defaults to 1. Results are stored in numbered subdirs
for each run and _ResultData array format is changed to include one
additional level of indexing used for iterations. For now we just
flatten the array whenever we use it but later submean statistical
analysis will use the information about the iterations.

Signed-off-by: Jan Kara <jack@suse.cz>