dependencies: install `bc` from run-mmtests.sh
As it's not always there, especially in minimal distro installs
(e.g., it's not there in a basic, no-gui, install of Tumbleweed).

Without it, `kernbench` fails like this:

FATAL kernbench: Failed to build vmlinux

And that's because trying building Linux fails like this:

  CHK     include/config/kernel.release
  CHK     include/generated/uapi/linux/version.h
  CHK     include/generated/utsrelease.h
  CHK     include/generated/bounds.h
  CHK     include/generated/timeconst.h
/bin/sh: bc: command not found
make[1]: *** [Kbuild:66: include/generated/timeconst.h] Error 127
make: *** [Makefile:986: prepare0] Error 2

In my experience, I've seen it needed only for `kernelbench`,
but it's small and generic enough, that I figured we could
just add it for everyone.

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>