mmtests: Add run-mmtests.sh --performance
This patch allows the user to force common cpu frequency related
performance settings, as opposed to doing it manually (or not at all).

I've added the obvious cpu governor 'performance' option, as well as
disabling turbo pstates and therefore preventing the cpu frequency to
rise further, and therefore potentially 'influence' results.

While performance parameters are perhaps not the common scenario in
production environments, it is a common way of both optimizing and
stabilizing benchmarking environments, particularly for those that
are cpu bound.

Signed-off-by: Davidlohr Bueso <dbueso@suse.de>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>