run-mmtests: make getopt accept -p as shortcut for --performance
Commit 2b346fd2a ("mmtests: Add run-mmtests.sh --performance) introduced
the new '--performance' command line argument. Looking at the argument
parsing code, it looks like '-p' would be accepted too. But it is not,
as it is not specified in getopt's short options list.

So add it there.

While there, update the usage() function so that all options are
actually mentioned.

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>