bonnie++: Add bonnie++ test used for bonnie version 1.98
Instead of updating bonnie I've decided to keep it as is (based on
version 1.03e) and use latest bonnie version for further development.
This allows easy comparison runs with old (bonnie) and new (bonnie++)
versions. After some grace period we can decide to remove the old
cruft entirely.

The initial version consists just of what we used so far -- only
bonnie-detailed-io-report.patch was ported to 1.98.

Signed-off-by: Andreas Herrmann <aherrmann@suse.com>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>