run-mmtests: keep monitors and management tasks out of CPU cgroup
If a CPU cgroup is created, it makes sense for it to contain only the
tasks that are actually performing the benchmark, not the management
mmtests processes, the monitors, etc.

In order to achieve that, move the CPU cgroup handling closer to
where the benchmark jobs are actually started and put inside it,
instead than the whole run-mmtests.sh and all its children, only
the tree of bin/run-single-test.sh.

Currently, CPU cgroups are used only to test a preliminary version
of Linux core-scheduling patches. In this specific case, having a
lot of tasks in the cgroup means that many more than just the ones
that are running the workload would need to be co-scheduled on
siblings (or not at all), potentially spoiling results.

I have verified that, with this change applied, the number of tasks
inside the CPU cgroup is sensibly lower. From the data I have,
performance don't change that much (most likely due to the fact
that monitors and management tasks sleep or do lightweight I/O
most of the time).

It still make sense to do this, though, as it makes the benchmark
look closer to how the real workload would be run (i.e., in this
specific case, only the tasks that the user wants to isolate from
the rest of the system will be co-scheduled/put inside the cgroup).

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>