run-kvm: fix cleanup of old mmtests done too early
We want to remove any trace of previous mmtests runs from inside the
guest, and we do so by deleting ~/git-private/$NAME, within the guest
itself. But we do it two times, the first one of which when $NAME has
not been defined and assigned a value yet!

Get rid of that.

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>