run-kvm: improve parameter handling
Make run-kvm.sh a bit more flexible when it comes to the parameters it
takes. We still avoid using `getopt` & friends, as it won't work well
with what will come in subsequent changes. We, however, make the (two,
currently) parameters the script takes "order independent". E.g., it's
no longer necessary for '--keep-kernel' to be the first one.

Note, however, that actual run-kvm.sh parameters still need to be put
before the parameters we want MMTests to use, when run inside the VM.

While there, document the above, and add a '-h' switch that prints some
usage.

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>