Gain some VM-lifecycle management capabilities
In order to be able to use MMTests as a standalone virtualization
benchmarking suite, we need for it to be able to deal with VMs lifecycle
management.

We do not focus on deploy, for now, but we at least need to be able to
start, stop and some more simple operations on VMs that exists, and are
known to libvirt, already.

Add some scripts for that.

More specifically, within SUSE, we use a tool called Marvin for handling
(among other things) the provisioning of hosts and VMs (see [1]). The
scripts added in this commit are taken from there. This is because,
internally, Marvin drives MMTests so, if the scripts are actually the
same:
- we minimize the need of changes (and the chance of breackages) for
  components, such as run-kvm.sh, which will have to work both from
  inside and from outside Marvin;
- any bugfix/improvement done, can be propagated to the original Marvin
  scripts (and vice versa).

In order, however, to make sure that Marvin uses its own copy, we put
them in a special directory, and add it at the end of the PATH. That way,
if Marvin has deployed the machine, versions in /usr/bin will be used;
if not, run-kvm.sh will use the MMTests specific ones.

We import the following scripts:
* kvm-ip-address
* kvm-start
* kvm-stop
* wait_ssh_available

[1] https://www.youtube.com/watch?v=jOnIQJQzW3s

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>