bin-virt/wait_ssh_available: fix missing return value propagation
The script, right now, return as an exit code whether or not the VM is
up and running and SSH is available only if the VM itself was running
already.

This is not an issue, only because we never check and/or rely on such
exit code to be actually representative of how the check went.

Fix this.

This change can be "ported" to Marvin copy of wait_ssh_available,
without causing any breakage.

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>