bin-virt/kvm-ip-address: take VM name as parameter
In order to make it possible for run-kvm.sh to work on VMs with
arbitrary names, the supporting scripts need to take the VM name
as a parameter.

Do this to kvm-ip-address. The new (`--vm foo`) parameter is optional
and, if provided, should be the first parameter.

This change can be "ported" to the Marvin copy of kvm-ip-address,
without causing any breakage.

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>