bin-virt/kvm-{start, stop}: take VM name as parameter
In order to make it possible for run-kvm.sh to work on VMs with
arbitrary names, the supporting scripts need to take the VM name
as a parameter.

Do this for kvm-start and kvm-stop. The new (`--vm foo`) parameter
is optional.

This change can be "ported" to the Marvin copies of kvm-{start,stop},
without causing any breakage.

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>