run-kvm: take VM name as parameter
Make it possible for run-kvm.sh to work on VMs with arbitrary name.
Now that all the supporting scripts are updated, we can add a
(optional) `--vm foo` parameter to the main program as well.

For making sure that the results of runs of the same benchmark in VMs
with different names are not overwritten, if the `--vm` parameter is
used we also slightly tweak the 'runname' (i.e., prefixing it with the
VM name).

This change will not break current Marvin usage of run-kvm.sh.

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>