bin-virt/kvm-{start, stop}: take a list of VM names as parameter
We not only want arbitrary names, we want to be able to run benchmarks
inside multiple VMs at the same time. In order to achieve this, make the
start and stop scripts able of acting on a list of VMs.

The `--vm foo` parameter can now be a list of VM names, and the scripts
iterate through such list and start (stop) all of them.

This, for now, happen sequentially, i.e., one VM is started (stopped) at
a time. And, each time, we wait until startup (shutdown) is completed.
This can be improved, i.e., multiple VMs can be started (stopped) in
parallel. This will be done in a later patch.

This change can be "ported" to the Marvin copies of kvm-{start,stop},
without causing any breakage.

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>